:recycle: [Refactor] App.tsx 분리 #main
