+++ packages/assignment/src/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./Complex";
+export * from "./ItemList";
+export * from "./NotificationSystem";
+export * from "./Header";
