+++ packages/assignment/src/context/ThemeContext.tsx
@@ -1,0 +1,31 @@
+import { createContext, useState, PropsWithChildren } from "react";
+import { useCallback, useMemo } from "../@lib";
+type Theme = "light" | "dark";
+
+interface ThemeContextType {
+  theme: Theme;
+  toggleTheme: () => void;
+}
+
+const ThemeContext = createContext<ThemeContextType | null>(null);
+
+const ThemeProvider: React.FC<PropsWithChildren> = ({ children }) => {
+  const [theme, setTheme] = useState<Theme>("light");
+  const toggleTheme = useCallback(() => {
+    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+  }, []);
+  const value = useMemo(() => ({ theme, toggleTheme }), [theme, toggleTheme]);
+  return (
+    <ThemeContext.Provider value={value}>
+      <div
+        className={`min-h-screen ${
+          theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"
+        }`}
+      >
+        {children}
+      </div>
+    </ThemeContext.Provider>
+  );
+};
+
+export { ThemeContext, ThemeProvider };
