+++ packages/assignment/src/context/UserContext.tsx
@@ -1,0 +1,37 @@
+import { createContext, useState, PropsWithChildren } from "react";
+import { useCallback, useMemo } from "../@lib";
+import { useNotification } from "../hooks";
+
+interface User {
+  id: number;
+  name: string;
+  email: string;
+}
+
+interface UserContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+const UserContext = createContext<UserContextType | null>(null);
+
+const UserProvider: React.FC<PropsWithChildren> = ({ children }) => {
+  const [user, setUser] = useState<User | null>(null);
+  const { addNotification } = useNotification();
+  const login = useCallback(
+    (email: string) => {
+      setUser({ id: 1, name: "홍길동", email });
+      addNotification("성공적으로 로그인되었습니다", "success");
+    },
+    [addNotification]
+  );
+  const logout = useCallback(() => {
+    setUser(null);
+    addNotification("성공적으로 로그아웃되었습니다", "success");
+  }, [addNotification]);
+  const value = useMemo(() => ({ user, login, logout }), [user, login, logout]);
+  return <UserContext.Provider value={value}>{children}</UserContext.Provider>;
+};
+
+export { UserContext, UserProvider };
